+++ gcs/CHANGES.md
@@ -121,6 +121,9 @@
 1.  Increment Hadoop `FileSystem.Statistics` counters for read and write
     operations.
 
+1.  Always infer implicit directories and remove
+    `fs.gs.implicit.dir.infer.enable` property.
+
 ### 2.1.1 - 2020-03-11
 
 1.  Add upload cache to support high-level retries of failed uploads. Cache size